requirements.txt
streamlit==1.32.2
opencv-python-headless==4.10.0.84
Pillow==10.3.0
numpy==1.26.4
torch==2.3.0
torchvision==0.18.0
git+https://github.com/facebookresearch/detectron2.git
